+++ libraries/DRV8825/test/unit_test_001.cpp
@@ -221,7 +221,8 @@ unittest(test_sleep)
 }
 
 
-
 unittest_main()
 
-// --------
+
+//  -- END OF FILE --
+