+++ libraries/DRV8825/CHANGELOG.md
@@ -6,11 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.4] - 2023-02-19
+- fix keywords.txt
+- add bool return value to enable, reset and sleep functions.
+  to indicate that these lines are defined or not.
+- update readme.md
+- update GitHub actions
+- update license 2023
+- minor edits
+
+
 ## [0.1.3] - 2022-11-02
 - add changelog.md
 - add rp2040 to build-CI
 - fix version number
-
 
 ## [0.1.2] - 2022-07-14
 - add support for SLP RST and EN pin